streamlit/requirements.txt
streamlit
pandas
snowflake-sqlalchemy
plotly